Moves remaining widgets to their module
